requirements.txt
vizdoom
gymnasium
stable-baselines3[extra]
torch
numpy
pyyaml
opencv-python
customtkinter